+++ src/front/js/component/navbar.js
@@ -7,7 +7,7 @@ export const Navbar = () => {
   const { store, actions } = useContext(Context);
   return (
     <nav className="navbar navbar-expand-lg  bg-danger p-1">
-      <Link to="/perfil">
+      <Link to="/photoperfil">
         <img
           id="imagenPerfilNavbar"
           className="img-fluid rounded-circle"
@@ -29,7 +29,7 @@ export const Navbar = () => {
       <div className="collapse navbar-collapse" id="navbarNavDropdown">
         <ul className="navbar-nav">
           <li className="nav-item active">
-            <Link to="/photoperfil">
+            <Link to="/home">
               <a className="nav-link" href="#">
                 Pagina principal <span className="sr-only">(current)</span>
               </a>
