+++ src/front/js/pages/home.js
@@ -1,4 +1,4 @@
-import React, { useContext } from "react";
+import React, { useContext, useEffect } from "react";
 import { Context } from "../store/appContext";
 import rigoImageUrl from "../../img/rigo-baby.jpg";
 import "../../styles/home.css";
@@ -6,7 +6,9 @@ import { Link } from "react-router-dom";
 
 export const Home = () => {
 	const { store, actions } = useContext(Context);
-
+	// useEffect(() => {
+	// 	actions.LoadImage()
+	// }, []);
 	return (
 		<div className="text-center mt-5">
 			<h1>Hello Rigo!!</h1>
