+++ src/front/js/store/flux.js
@@ -3,6 +3,7 @@ const getState = ({ getStore, getActions, setStore }) => {
     store: {
       token: "",
       imagen: "https://img.freepik.com/vector-premium/perfil-hombre-dibujos-animados_18591-58482.jpg?w=200",
+      respuesta: "",
     },
     actions: {
       logout: () => {
@@ -11,6 +12,7 @@ const getState = ({ getStore, getActions, setStore }) => {
       },
 
       login: (email, password) => {
+        const actions = getActions()
         fetch(process.env.BACKEND_URL + "/api/token", {
           method: "POST",
           headers: {
@@ -31,9 +33,51 @@ const getState = ({ getStore, getActions, setStore }) => {
             }
           })
           .then((respuestajson) => {
+            actions.Load(respuestajson.access_token)
             sessionStorage.setItem("token", respuestajson.access_token);
             setStore({ token: respuestajson.access_token });
           });
+      },
+      LoadImage: (evt) => {
+        const store = getStore();
+        const options = {
+          headers: {
+            "Content-Type": "application/json",
+            Accept: "application/json",
+            Authorization: "Bearer " + store.token,
+          },
+          method: "GET",
+        }
+        fetch(process.env.BACKEND_URL + "/api/load", options)
+          .then(respuestadelback =>
+            respuestadelback.json())
+          .then(data => {
+            setStore({ imagen: data })
+            setStore({ respuesta: "" })
+          })
+
+      },
+      Load: (parametro) => {
+        const options = {
+          headers: {
+            "Content-Type": "application/json",
+            Accept: "application/json",
+            Authorization: "Bearer " + parametro
+          },
+          method: "GET",
+        }
+        fetch(process.env.BACKEND_URL + "/api/load", options)
+          .then(respuestadelback =>
+            respuestadelback.json())
+          .then(data => {
+            setStore({ imagen: data })
+            setStore({ respuesta: "" })
+          })
+
+      },
+
+      getrespuesta: (str) => {
+        setStore({ respuesta: str })
       },
 
       //FUNCION reloadToken PARA QUE NO SE PIERDA EL TOKEN DEL STORAGE
